Refactored single and multi distribution forms to their own files.  Added validation handling for the multi distribution percentages.  Tweak as needed/desired
