+++ src/types/Distribution.ts
@@ -3,6 +3,10 @@ export interface IDistributionBase {
   rollout: number;
 }
 
+export interface IDistributionVariant extends IDistributionBase {
+  variantKey: string;
+}
+
 export interface IDistribution extends IDistributionBase {
   id: string;
   createdAt: string;
