+++ src/types/Variant.ts
@@ -1,3 +1,5 @@
+import { ISelectable } from '~/components/forms/Combobox';
+
 export interface IVariantBase {
   key: string;
   name: string;
@@ -10,3 +12,5 @@ export interface IVariant extends IVariantBase {
   createdAt: string;
   updatedAt: string;
 }
+
+export type SelectableVariant = IVariant & ISelectable;
